js: Define and document the end-user API
Adds an API that wraps AesSiv, for the event when we add AES-GCM-SIV support in
the future.

Adds an initial README.md for the JavaScript subproject which documents the API.
