+++ js/test/util.spec.ts
@@ -1,6 +1,6 @@
 import { suite, test } from "mocha-typescript";
 import { expect } from "chai";
-import { wipe } from "../src/internal/wipe";
+import { wipe } from "../src/internal/util";
 
 @suite class WipeSpec {
   @test "should wipe bytes"() {
